refactoring code & structure files
